requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=1.4.0
mysql-connector-python>=8.0
